Finish converting PdfIntegrationTests to PdfIntegrationMockMvcTest
[#178367884]
